refactored stores
