+++ src/renderer/store/index.ts
@@ -1,3 +1,4 @@
+import { Reducer } from 'redux';
 import * as Layout from './layout';
 
 // The top-level state object
@@ -8,7 +9,7 @@ export interface ApplicationState {
 // Whenever an action is dispatched, Redux will update each top-level application state property
 // using the reducer with the matching name. It's important that the names match exactly, and that
 // the reducer acts on the corresponding ApplicationState property type.
-export const reducers = {
+export const reducers: {[key: string]: Reducer<any>} = {
   layout: Layout.reducer
 };
