--- src/renderer/store/layout/layout.actions.ts
@@ -1,12 +1,0 @@
-import { ActionCreator } from 'redux';
-import { SetWindowMaximizedStateAction } from './layout.types';
-
-const setWindowMaximizedState: ActionCreator<SetWindowMaximizedStateAction> =
-  (payload: boolean) => ({
-    payload,
-    type: '@@layout/SET_WINDOW_MAXIMIZED_STATE'
-  });
-
-export const actionCreators = {
-  setWindowMaximizedState
-};
