+++ src/renderer/store/layout/types.ts
